@@ -10,11 +10,13 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": "^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
-		"test": "xo && ava && tsd",
+		"test": "xo && ava --serial --no-worker-threads && tsd",
 		"bench": "node benchmark.js"
 	},
 	"files": [
@@ -56,11 +58,11 @@
 		"slash": "^3.0.0"
 	},
 	"devDependencies": {
-		"ava": "^2.4.0",
+		"ava": "^4.3.1",
 		"benchmark": "^2.1.4",
 		"make-dir": "^3.1.0",
 		"tempy": "^0.7.0",
-		"tsd": "^0.13.1",
-		"xo": "^0.33.1"
+		"tsd": "^0.22.0",
+		"xo": "^0.50.0"
 	}
 }
