@@ -1,25 +1,25 @@
 import {expectType} from 'tsd';
-import del = require('.');
+import {deleteAsync, deleteSync} from './index.js';
 
 const paths = [
 	'temp/*.js',
-	'!temp/unicorn.js'
+	'!temp/unicorn.js',
 ];
 
 // Del
-expectType<Promise<string[]>>(del('temp/*.js'));
-expectType<Promise<string[]>>(del(paths));
+expectType<Promise<string[]>>(deleteAsync('temp/*.js'));
+expectType<Promise<string[]>>(deleteAsync(paths));
 
-expectType<Promise<string[]>>(del(paths, {force: true}));
-expectType<Promise<string[]>>(del(paths, {dryRun: true}));
-expectType<Promise<string[]>>(del(paths, {concurrency: 20}));
-expectType<Promise<string[]>>(del(paths, {cwd: ''}));
+expectType<Promise<string[]>>(deleteAsync(paths, {force: true}));
+expectType<Promise<string[]>>(deleteAsync(paths, {dryRun: true}));
+expectType<Promise<string[]>>(deleteAsync(paths, {concurrency: 20}));
+expectType<Promise<string[]>>(deleteAsync(paths, {cwd: ''}));
 
 // Del (sync)
-expectType<string[]>(del.sync('tmp/*.js'));
-expectType<string[]>(del.sync(paths));
+expectType<string[]>(deleteSync('tmp/*.js'));
+expectType<string[]>(deleteSync(paths));
 
-expectType<string[]>(del.sync(paths, {force: true}));
-expectType<string[]>(del.sync(paths, {dryRun: true}));
-expectType<string[]>(del.sync(paths, {concurrency: 20}));
-expectType<string[]>(del.sync(paths, {cwd: ''}));
+expectType<string[]>(deleteSync(paths, {force: true}));
+expectType<string[]>(deleteSync(paths, {dryRun: true}));
+expectType<string[]>(deleteSync(paths, {concurrency: 20}));
+expectType<string[]>(deleteSync(paths, {cwd: ''}));
