@@ -1,10 +1,13 @@
-import path from 'path';
-import fs from 'fs';
-import {serial as test} from 'ava';
+import {fileURLToPath} from 'node:url';
+import fs from 'node:fs';
+import path from 'node:path';
+import process from 'node:process';
+import test from 'ava';
 import tempy from 'tempy';
 import makeDir from 'make-dir';
-import del from '.';
+import {deleteAsync, deleteSync} from './index.js';
 
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const processCwd = process.cwd();
 
 function exists(t, files) {
@@ -24,7 +27,7 @@ const fixtures = [
 	'2.tmp',
 	'3.tmp',
 	'4.tmp',
-	'.dot.tmp'
+	'.dot.tmp',
 ];
 
 test.beforeEach(t => {
@@ -36,23 +39,23 @@ test.beforeEach(t => {
 });
 
 test('delete files - async', async t => {
-	await del(['*.tmp', '!1*'], {cwd: t.context.tmp});
+	await deleteAsync(['*.tmp', '!1*'], {cwd: t.context.tmp});
 
 	exists(t, ['1.tmp', '.dot.tmp']);
 	notExists(t, ['2.tmp', '3.tmp', '4.tmp']);
 });
 
 test('delete files - sync', t => {
-	del.sync(['*.tmp', '!1*'], {cwd: t.context.tmp});
+	deleteSync(['*.tmp', '!1*'], {cwd: t.context.tmp});
 
 	exists(t, ['1.tmp', '.dot.tmp']);
 	notExists(t, ['2.tmp', '3.tmp', '4.tmp']);
 });
 
 test('take options into account - async', async t => {
-	await del(['*.tmp', '!1*'], {
+	await deleteAsync(['*.tmp', '!1*'], {
 		cwd: t.context.tmp,
-		dot: true
+		dot: true,
 	});
 
 	exists(t, ['1.tmp']);
@@ -60,9 +63,9 @@ test('take options into account - async', async t => {
 });
 
 test('take options into account - sync', t => {
-	del.sync(['*.tmp', '!1*'], {
+	deleteSync(['*.tmp', '!1*'], {
 		cwd: t.context.tmp,
-		dot: true
+		dot: true,
 	});
 
 	exists(t, ['1.tmp']);
@@ -71,50 +74,50 @@ test('take options into account - sync', t => {
 
 test('return deleted files - async', async t => {
 	t.deepEqual(
-		await del('1.tmp', {cwd: t.context.tmp}),
-		[path.join(t.context.tmp, '1.tmp')]
+		await deleteAsync('1.tmp', {cwd: t.context.tmp}),
+		[path.join(t.context.tmp, '1.tmp')],
 	);
 });
 
 test('return deleted files - sync', t => {
 	t.deepEqual(
-		del.sync('1.tmp', {cwd: t.context.tmp}),
-		[path.join(t.context.tmp, '1.tmp')]
+		deleteSync('1.tmp', {cwd: t.context.tmp}),
+		[path.join(t.context.tmp, '1.tmp')],
 	);
 });
 
 test('don\'t delete files, but return them - async', async t => {
-	const deletedFiles = await del(['*.tmp', '!1*'], {
+	const deletedFiles = await deleteAsync(['*.tmp', '!1*'], {
 		cwd: t.context.tmp,
-		dryRun: true
+		dryRun: true,
 	});
 	exists(t, ['1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
 	t.deepEqual(deletedFiles, [
 		path.join(t.context.tmp, '2.tmp'),
 		path.join(t.context.tmp, '3.tmp'),
-		path.join(t.context.tmp, '4.tmp')
+		path.join(t.context.tmp, '4.tmp'),
 	]);
 });
 
 test('don\'t delete files, but return them - sync', t => {
-	const deletedFiles = del.sync(['*.tmp', '!1*'], {
+	const deletedFiles = deleteSync(['*.tmp', '!1*'], {
 		cwd: t.context.tmp,
-		dryRun: true
+		dryRun: true,
 	});
 	exists(t, ['1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
 	t.deepEqual(deletedFiles, [
 		path.join(t.context.tmp, '2.tmp'),
 		path.join(t.context.tmp, '3.tmp'),
-		path.join(t.context.tmp, '4.tmp')
+		path.join(t.context.tmp, '4.tmp'),
 	]);
 });
 
 // Currently this is only testable locally on macOS.
 // https://github.com/sindresorhus/del/issues/68
 test('does not throw EINVAL - async', async t => {
-	await del('**/*', {
+	await deleteAsync('**/*', {
 		cwd: t.context.tmp,
-		dot: true
+		dot: true,
 	});
 
 	const nestedFile = path.resolve(t.context.tmp, 'a/b/c/nested.js');
@@ -125,16 +128,16 @@ test('does not throw EINVAL - async', async t => {
 		makeDir.sync(nestedFile);
 
 		// eslint-disable-next-line no-await-in-loop
-		const removed = await del('**/*', {
+		const removed = await deleteAsync('**/*', {
 			cwd: t.context.tmp,
-			dot: true
+			dot: true,
 		});
 
 		const expected = [
 			path.resolve(t.context.tmp, 'a'),
 			path.resolve(t.context.tmp, 'a/b'),
 			path.resolve(t.context.tmp, 'a/b/c'),
-			path.resolve(t.context.tmp, 'a/b/c/nested.js')
+			path.resolve(t.context.tmp, 'a/b/c/nested.js'),
 		];
 
 		t.deepEqual(removed, expected);
@@ -147,9 +150,9 @@ test('does not throw EINVAL - async', async t => {
 });
 
 test('does not throw EINVAL - sync', t => {
-	del.sync('**/*', {
+	deleteSync('**/*', {
 		cwd: t.context.tmp,
-		dot: true
+		dot: true,
 	});
 
 	const nestedFile = path.resolve(t.context.tmp, 'a/b/c/nested.js');
@@ -159,16 +162,16 @@ test('does not throw EINVAL - sync', t => {
 	while (count !== totalAttempts) {
 		makeDir.sync(nestedFile);
 
-		const removed = del.sync('**/*', {
+		const removed = deleteSync('**/*', {
 			cwd: t.context.tmp,
-			dot: true
+			dot: true,
 		});
 
 		const expected = [
 			path.resolve(t.context.tmp, 'a'),
 			path.resolve(t.context.tmp, 'a/b'),
 			path.resolve(t.context.tmp, 'a/b/c'),
-			path.resolve(t.context.tmp, 'a/b/c/nested.js')
+			path.resolve(t.context.tmp, 'a/b/c/nested.js'),
 		];
 
 		t.deepEqual(removed, expected);
@@ -181,14 +184,14 @@ test('does not throw EINVAL - sync', t => {
 });
 
 test('delete relative files outside of process.cwd using cwd - async', async t => {
-	await del(['1.tmp'], {cwd: t.context.tmp});
+	await deleteAsync(['1.tmp'], {cwd: t.context.tmp});
 
 	exists(t, ['2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
 	notExists(t, ['1.tmp']);
 });
 
 test('delete relative files outside of process.cwd using cwd - sync', t => {
-	del.sync(['1.tmp'], {cwd: t.context.tmp});
+	deleteSync(['1.tmp'], {cwd: t.context.tmp});
 
 	exists(t, ['2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
 	notExists(t, ['1.tmp']);
@@ -196,7 +199,7 @@ test('delete relative files outside of process.cwd using cwd - sync', t => {
 
 test('delete absolute files outside of process.cwd using cwd - async', async t => {
 	const absolutePath = path.resolve(t.context.tmp, '1.tmp');
-	await del([absolutePath], {cwd: t.context.tmp});
+	await deleteAsync([absolutePath], {cwd: t.context.tmp});
 
 	exists(t, ['2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
 	notExists(t, ['1.tmp']);
@@ -204,7 +207,7 @@ test('delete absolute files outside of process.cwd using cwd - async', async t =
 
 test('delete absolute files outside of process.cwd using cwd - sync', t => {
 	const absolutePath = path.resolve(t.context.tmp, '1.tmp');
-	del.sync([absolutePath], {cwd: t.context.tmp});
+	deleteSync([absolutePath], {cwd: t.context.tmp});
 
 	exists(t, ['2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
 	notExists(t, ['1.tmp']);
@@ -213,9 +216,9 @@ test('delete absolute files outside of process.cwd using cwd - sync', t => {
 test('cannot delete actual working directory without force: true - async', async t => {
 	process.chdir(t.context.tmp);
 
-	await t.throwsAsync(del([t.context.tmp]), {
+	await t.throwsAsync(deleteAsync([t.context.tmp]), {
 		instanceOf: Error,
-		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['', '1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -226,10 +229,10 @@ test('cannot delete actual working directory without force: true - sync', t => {
 	process.chdir(t.context.tmp);
 
 	t.throws(() => {
-		del.sync([t.context.tmp]);
+		deleteSync([t.context.tmp]);
 	}, {
 		instanceOf: Error,
-		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['', '1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -239,9 +242,9 @@ test('cannot delete actual working directory without force: true - sync', t => {
 test('cannot delete actual working directory with cwd option without force: true - async', async t => {
 	process.chdir(t.context.tmp);
 
-	await t.throwsAsync(del([t.context.tmp], {cwd: __dirname}), {
+	await t.throwsAsync(deleteAsync([t.context.tmp], {cwd: __dirname}), {
 		instanceOf: Error,
-		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['', '1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -252,10 +255,10 @@ test('cannot delete actual working directory with cwd option without force: true
 	process.chdir(t.context.tmp);
 
 	t.throws(() => {
-		del.sync([t.context.tmp], {cwd: __dirname});
+		deleteSync([t.context.tmp], {cwd: __dirname});
 	}, {
 		instanceOf: Error,
-		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['', '1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -265,9 +268,9 @@ test('cannot delete actual working directory with cwd option without force: true
 test('cannot delete files outside cwd without force: true - async', async t => {
 	const absolutePath = path.resolve(t.context.tmp, '1.tmp');
 
-	await t.throwsAsync(del([absolutePath]), {
+	await t.throwsAsync(deleteAsync([absolutePath]), {
 		instanceOf: Error,
-		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -277,10 +280,10 @@ test('cannot delete files outside cwd without force: true - sync', t => {
 	const absolutePath = path.resolve(t.context.tmp, '1.tmp');
 
 	t.throws(() => {
-		del.sync([absolutePath]);
+		deleteSync([absolutePath]);
 	}, {
 		instanceOf: Error,
-		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['', '1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -291,9 +294,9 @@ test('cannot delete files inside process.cwd when outside cwd without force: tru
 	const removeFile = path.resolve(t.context.tmp, '2.tmp');
 	const cwd = path.resolve(t.context.tmp, '1.tmp');
 
-	await t.throwsAsync(del([removeFile], {cwd}), {
+	await t.throwsAsync(deleteAsync([removeFile], {cwd}), {
 		instanceOf: Error,
-		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -306,10 +309,10 @@ test('cannot delete files inside process.cwd when outside cwd without force: tru
 	const cwd = path.resolve(t.context.tmp, '1.tmp');
 
 	t.throws(() => {
-		del.sync([removeFile], {cwd});
+		deleteSync([removeFile], {cwd});
 	}, {
 		instanceOf: Error,
-		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.'
+		message: 'Cannot delete files/directories outside the current working directory. Can be overridden with the `force` option.',
 	});
 
 	exists(t, ['1.tmp', '2.tmp', '3.tmp', '4.tmp', '.dot.tmp']);
@@ -319,7 +322,7 @@ test('cannot delete files inside process.cwd when outside cwd without force: tru
 test('windows can pass absolute paths with "\\" - async', async t => {
 	const filePath = path.resolve(t.context.tmp, '1.tmp');
 
-	const removeFiles = await del([filePath], {cwd: t.context.tmp, dryRun: true});
+	const removeFiles = await deleteAsync([filePath], {cwd: t.context.tmp, dryRun: true});
 
 	t.deepEqual(removeFiles, [filePath]);
 });
@@ -327,7 +330,7 @@ test('windows can pass absolute paths with "\\" - async', async t => {
 test('windows can pass absolute paths with "\\" - sync', t => {
 	const filePath = path.resolve(t.context.tmp, '1.tmp');
 
-	const removeFiles = del.sync([filePath], {cwd: t.context.tmp, dryRun: true});
+	const removeFiles = deleteSync([filePath], {cwd: t.context.tmp, dryRun: true});
 
 	t.deepEqual(removeFiles, [filePath]);
 });
@@ -336,7 +339,7 @@ test('windows can pass relative paths with "\\" - async', async t => {
 	const nestedFile = path.resolve(t.context.tmp, 'a/b/c/nested.js');
 	makeDir.sync(nestedFile);
 
-	const removeFiles = await del([nestedFile], {cwd: t.context.tmp, dryRun: true});
+	const removeFiles = await deleteAsync([nestedFile], {cwd: t.context.tmp, dryRun: true});
 
 	t.deepEqual(removeFiles, [nestedFile]);
 });
@@ -345,7 +348,7 @@ test('windows can pass relative paths with "\\" - sync', t => {
 	const nestedFile = path.resolve(t.context.tmp, 'a/b/c/nested.js');
 	makeDir.sync(nestedFile);
 
-	const removeFiles = del.sync([nestedFile], {cwd: t.context.tmp, dryRun: true});
+	const removeFiles = deleteSync([nestedFile], {cwd: t.context.tmp, dryRun: true});
 
 	t.deepEqual(removeFiles, [nestedFile]);
 });
@@ -353,28 +356,28 @@ test('windows can pass relative paths with "\\" - sync', t => {
 test('onProgress option - progress of non-existent file', async t => {
 	let report;
 
-	await del('non-existent-directory', {onProgress: event => {
+	await deleteAsync('non-existent-directory', {onProgress(event) {
 		report = event;
 	}});
 
 	t.deepEqual(report, {
 		totalCount: 0,
 		deletedCount: 0,
-		percent: 1
+		percent: 1,
 	});
 });
 
 test('onProgress option - progress of single file', async t => {
 	let report;
 
-	await del(t.context.tmp, {cwd: __dirname, force: true, onProgress: event => {
+	await deleteAsync(t.context.tmp, {cwd: __dirname, force: true, onProgress(event) {
 		report = event;
 	}});
 
 	t.deepEqual(report, {
 		totalCount: 1,
 		deletedCount: 1,
-		percent: 1
+		percent: 1,
 	});
 });
 
@@ -383,17 +386,17 @@ test('onProgress option - progress of multiple files', async t => {
 
 	const sourcePath = process.platform === 'win32' ? path.resolve(`${t.context.tmp}/*`).replace(/\\/g, '/') : `${t.context.tmp}/*`;
 
-	await del(sourcePath, {
+	await deleteAsync(sourcePath, {
 		cwd: __dirname,
 		force: true,
-		onProgress: event => {
+		onProgress(event) {
 			report = event;
-		}
+		},
 	});
 
 	t.deepEqual(report, {
 		totalCount: 4,
 		deletedCount: 4,
-		percent: 1
+		percent: 1,
 	});
 });
